+++ utils/ftqq.py
@@ -2,7 +2,7 @@
 # @Author       : Chr_
 # @Date         : 2020-07-29 14:08:11
 # @LastEditors  : Chr_
-# @LastEditTime : 2020-07-30 16:44:48
+# @LastEditTime : 2020-08-01 22:30:29
 # @Description  : 方糖气球模块
 '''
 
@@ -13,16 +13,16 @@ from json import JSONDecodeError
 logger = get_logger('FTQQ')
 
 
-def send_to_ftqq(title: str, text: str, skey: str) -> bool:
+def send_to_ftqq(title: str, text: str, ftqqcfg: dict) -> bool:
     '''发送消息到方糖气球
     参数:
         title: 标题
         text: 内容
-        skey: 方糖气球SKEY
+        ftqqcfg: 方糖气球配置节
     返回:
-        bool: 是否发送成功
+        bool: 是否成功
     '''
-    url = f'https://sc.ftqq.com/{skey}.send'
+    url = f'https://sc.ftqq.com/{ftqqcfg.get("skey")}.send'
     data = {'text': str(title), 'desp': str(text)}
     resp = requests.post(url=url, data=data)
     try: